requirements.txt
fastapi==0.111.0
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
pydantic==2.7.1
pytest==8.2.1
segmentation_models_pytorch==0.3.3
torch==2.3.0
torchvision==0.18.0
uvicorn==0.29.0
